fix(ui): stop 32-bit truncation of composition cell ids in k-NN scratch
Galaxy-composition Cells carry 2^52-range ids; the pre-allocated
Int32Array scratch buffer wrapped them into phantom node ids, crashing
the topology worker ('can't access property pos_seed') and forcing the
synchronous main-thread fallback build (multi-second canvas freeze,
stale display graph, unclickable cells during inspection).

+++ packages/ui/src/geometry/neighborGraph.ts
@@ -152,9 +152,10 @@ export function buildNeighborGraph(
 
   // Pre-allocated parallel scratch buffers for top-k. Avoids the
   // per-iteration object alloc that the previous implementation hit
-  // ~N²/k times.
+  // ~N²/k times. Ids must be Float64: galaxy-composition Cells carry
+  // 2^52-range ids that an Int32Array would wrap into phantom nodes.
   const SCRATCH_CAP = Math.max(64, k * 16);
-  const scratchId = new Int32Array(SCRATCH_CAP);
+  const scratchId = new Float64Array(SCRATCH_CAP);
   const scratchDSq = new Float32Array(SCRATCH_CAP);
 
   // 2D spatial-hash bucketing cells by xz coords. k-NN scans only the
